requirements.txt
ffmpeg-python
numpy==1.19.3 # OpenCV 4.5.1 pip requires numpy >= 1.19.3
waggle[all] @ https://github.com/waggle-sensor/pywaggle/releases/download/0.46.2/waggle-0.46.2-py3-none-any.whl
